+++ tools/src/check-packages/checkDependenciesAsync.ts
@@ -33,10 +33,7 @@ type SourceFileImportRef = {
   isTypeOnly?: boolean;
 };
 
-// We are incrementally rolling this out, the sdk packages in this list are expected to be invalid
-const IGNORED_PACKAGES = [
-  'expo-updates', // cli: @expo/plist, debug, getenv - utils: @expo/cli, @expo/metro-config, metro
-];
+const IGNORED_PACKAGES: string[] = [];
 
 const SPECIAL_DEPENDENCIES: Record<string, Record<string, IgnoreKind | void> | void> = {
   'expo-dev-menu': {
@@ -83,7 +80,6 @@ const SPECIAL_DEPENDENCIES: Record<string, Record<string, IgnoreKind | void> | v
     '@expo/metro-config/build/babel-transformer': 'types-only',
     'react-native-worklets/plugin': 'ignore-dev', // Checked via hasModule before requiring
     'react-native-reanimated/plugin': 'ignore-dev', // Checked via hasModule before requiring
-    'expo/config': 'ignore-dev', // WARN: May need a reverse peer dependency
   },
 };
 
@@ -131,7 +127,7 @@ export async function checkDependenciesAsync(pkg: Package, type: PackageCheckTyp
     for (const importRef of source.importRefs) {
       if (importRef.type !== 'external' || pkg.packageName === importRef.packageName) {
         continue;
-      } else if (isDisallowedImport(importRef, pkg.packageName)) {
+      } else if (isDisallowedImport(importRef)) {
         invalidImports.push({ file: source.file, importRef, kind: undefined });
       } else if (isIgnoredPackage) {
         continue;
@@ -183,7 +179,7 @@ export async function checkDependenciesAsync(pkg: Package, type: PackageCheckTyp
       Logger.verbose(
         `     > ${path.relative(pkg.path, file.path)} - ${importRef.importValue}` +
           `${importRef.isTypeOnly ? ' (types only)' : ''}` +
-          `${isDisallowedImport(importRef, pkg.packageName) ? ' (disallowed)' : ''}`
+          `${isDisallowedImport(importRef) ? ' (disallowed)' : ''}`
       );
     });
 
@@ -198,13 +194,7 @@ function isNCCBuilt(pkg: Package): boolean {
   return !!pkg.packageJson.bin && !!buildScript?.includes('ncc');
 }
 
-function isDisallowedImport(ref: SourceFileImportRef, sourceName: string): boolean {
-  if (sourceName === 'expo-updates') {
-    // TODO: We're currently allowing disallowed `metro` imports in `expo-updates`
-    // since the same file that contains them also contains more invalid imports
-    // that should all be fixed in one go
-    return false;
-  }
+function isDisallowedImport(ref: SourceFileImportRef): boolean {
   const packageName = getPackageName(ref.packageName);
   return packageName === 'metro' || packageName.startsWith('metro-');
 }